Split off config reader to its own module so we can reuse it for sbotcli
